feat: initial transition ans splash screen added
